+++ src/main/java/chess/dao/MovementDao.java
@@ -1,37 +1,14 @@
 package chess.dao;
 
-import chess.database.JdbcTemplate;
 import chess.dto.MovementDto;
 import java.util.List;
 
-public class MovementDao {
+public interface MovementDao {
 
-    private final JdbcTemplate jdbcTemplate;
-
-    public MovementDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public void addMovement(MovementDto movementDto) {
-        jdbcTemplate.update(
-                "INSERT INTO movement(turn, source, target)VALUES(?, ?, ?)",
-                movementDto.turn(),
-                movementDto.source(),
-                movementDto.target());
-    }
+    void addMovement(MovementDto movementDto);
 
 
-    public List<MovementDto> findAll() {
-        return jdbcTemplate.query(
-                "SELECT * FROM movement",
-                resultSet -> new MovementDto(
-                        resultSet.getString("turn"),
-                        resultSet.getString("source"),
-                        resultSet.getString("target")));
-    }
+    List<MovementDto> findAll();
 
-    public void deleteAll() {
-        jdbcTemplate.update(
-                "DELETE FROM movement");
-    }
+    void deleteAll();
 }
